Refactor(Factory) : added instance of factory
